Added JS logic
Added the logic so the user can press one of the buttons on their
keyboard, that correspond with what's displayed on the screen. When
the user presses a key down, the drum is highlighted and the sound
is played. When the user releases the button, the highlighted
instrument is no longer highlighted.